reporting/xas/requirements.txt
python-dotenv
flask
gunicorn
gevent
flask_cors
elasticsearch>=6.0.0,<7.0.0
pytest
pyjwt
python-jose